A good time for a commit ;)
Wow, finally I'm happy with the rendering result on my Notebook ;D. And I'll have to vommit soon since i played too much with the animations
Lots has been done and a major handbreak has been lifted.
The cammera created by von-grid code had a far-clipping plane of 10'000 which is fine if you don't want to grow your
map according to frustum-analysis. Since frustum was way too big, I introduced a workaround which was  first checking if distance is inside a specified radius of the cammera.
Changing the far-clipping plane made this workaround (and the resulting side-effects) uneccessary.
Even fog applies to the end of the clipping-plane now ;)

other improvements:

* naming improved
* config extendes (almost everything now configurable there)
* performance is good on my notebook (phone can be improved)
* adds animations ;)
* updates view less often (still optimizable)
* debug-HUD improved
* requests get served immediately (if meshes available)